+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] - 2019-06-21
+### Added 
+- different strategies for rounding last entry remaining time with new `-remainig` argument
+- corrected way all arguments works
+
 ## [0.0.4] - 2019-06-20
 ### Added
 - correct handling of overlapping entries